# Generic KV backend based on Bolt2