--- a/graph/kv/all_iterator.go
+++ b/graph/kv/all_iterator.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package bolt2
+package kv
 
 import (
 	"github.com/cayleygraph/cayley/graph"