--- a/graph/kv/indexing_test.go
+++ b/graph/kv/indexing_test.go
@@ -1,4 +1,4 @@
-package bolt2
+package kv
 
 import "testing"