--- a/graph/kv/iterators.go
+++ b/graph/kv/iterators.go
@@ -1,4 +1,4 @@
-package bolt2
+package kv
 
 import (
 	"github.com/cayleygraph/cayley/graph"