--- a/graph/kv/quadstore_iterator_optimize.go
+++ b/graph/kv/quadstore_iterator_optimize.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package bolt2
+package kv
 
 import (
 	"github.com/cayleygraph/cayley/graph"