Make a generic KV backend based on Bolt.

Benchmarks shows only a random noise, so at least it does not hurts performance.
```
benchmark                                     old ns/op     new ns/op     delta
BenchmarkNamePredicate-8                      123620        121928        -1.37%
BenchmarkLargeSetsNoIntersection-8            11546478      11193748      -3.05%
BenchmarkVeryLargeSetsSmallIntersection-8     20313099      19839123      -2.33%
BenchmarkHelplessContainsChecker-8            703804251     670880049     -4.68%
BenchmarkHelplessNotContainsFilms-8           703763546     678679708     -3.56%
BenchmarkHelplessNotContainsActors-8          699743599     685286832     -2.07%
BenchmarkNetAndSpeed-8                        1139628       1128880       -0.94%
BenchmarkKeanuAndNet-8                        645955        636241        -1.50%
BenchmarkKeanuAndSpeed-8                      937512        932961        -0.49%
BenchmarkKeanuOther-8                         5464889       5429001       -0.66%
BenchmarkKeanuBullockOther-8                  10019729      9762838       -2.56%
BenchmarkSaveBogartPerformances-8             315197        315232        +0.01%
```